refactor(budget): organize internal structure and unit tests
