+++ spring-ai-ledger-budget/src/main/java/io/springaileger/budget/BudgetState.java
@@ -1,11 +1,11 @@
 package io.springaileger.budget;
 
 /**
- * BudgetState는 "이번 AI 호출을 허용할지 말지"에 대한 최종 판단 결과이다.
+ * 예산 평가 결과 상태를 나타냅니다.
  *
- * - ALLOW : 예산에 여유가 있어서 정상 호출 가능
- * - WARN  : 예산이 거의 찼지만, 이번 호출은 허용
- * - BLOCK : 예산을 초과했기 때문에 호출을 차단
+ * ALLOW : 호출 허용
+ * WARN  : 예산 경고
+ * BLOCK : 호출 차단
  */
 
 public enum BudgetState {
